refactor(plugins): move examples/plugins to plugins/, add marketplace manifest
The repo is now itself a plugin marketplace: worca-cc-marketplace.json at the
root declares the 5 bundled plugins. examples/ is gone; sandbox (smoke) moves
to /sandbox.
