+++ plugins/discord-chat/channel/worker.mjs
