+++ plugins/discord-chat/lib/markdown.mjs
