+++ plugins/discord-chat/lib/message.mjs
