+++ plugins/discord-chat/lib/segments.mjs
