+++ plugins/discord-chat/worca-cc-plugin.json
