+++ plugins/slack-chat/worca-cc-plugin.json
