+++ plugins/teams-chat/lib/message.mjs
