+++ plugins/teams-chat/lib/segments.mjs
