+++ plugins/teams-chat/lib/send-util.mjs
