+++ plugins/teams-chat/worca-cc-plugin.json
