+++ plugins/telegram-chat/lib/segments.mjs
