+++ plugins/telegram-chat/worca-cc-plugin.json
